Integrate x64 mode squashing into the type promotion lattice.
Currently we perform type promotion ignoring the state of the x64-mode flag, and canonicalize/squash 64-bit dtypes after the performing promotion. In the future, we wish to allow x64 types to be present and participate in type promotion even if we aren't in an x64-enabled mode, and this means that we need to avoid using canonicalization here. Instead, teach the promotion lattice about x64 mode directly.

We have the following intuition: in non-x64 mode, type promotion should never introduce an x64 type if one was not already present. This means that if all the types that are input to a promotion are squashed, then promotion will not introduce new x64 types.

It turns out to be fairly easy to do this. Normally we have the following type promotion lattice:

```mermaid
graph LR
        direction LR
        b1 --> i*
        i* --> i1
        i* --> u1
        u1 --> u2 --> u4 --> u8
        i1 --> i2 --> i4 --> i8
        u1 --> i2
        u2 --> i4
        u4 --> i8
        i8 --> f*
        u8 --> f*
        f* --> c*
        f* --> f2
        f* --> bf
        bf --> f4
        f2 --> f4
        c* --> c8
        f4 --> c8
        f4 --> f8
        c8 --> c16
        f8 --> c16
```

If x64 mode is disabled, we use the following modified lattice, where i4 and u4 promote to i4 instead of i8.

```mermaid
graph LR
        direction LR
        b1 --> i*
        i* --> i1
        i* --> u1
        u1 --> u2 --> u4 --> u8
        i1 --> i2 --> i4 --> i8
        u1 --> i2
        u2 --> i4
        u4 --> i4
        i8 --> f*
        u8 --> f*
        f* --> c*
        f* --> f2
        f* --> bf
        bf --> f4
        f2 --> f4
        c* --> c8
        f4 --> c8
        f4 --> f8
        c8 --> c16
        f8 --> c16
```

We can then remove the type canonicalization on the output of promotion.

PiperOrigin-RevId: 800205545